Split home page into components
