+++ components/Header/styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Header = styled('h1')`
+  background-color: rgba(255, 255, 255, 0.6);
+  border-bottom: 1px solid rgba(0, 0, 0, 0.06);
+  color: #4e4a49;
+  font-size: 1.375rem;
+  font-weight: bold;
+  line-height: 1rem;
+  margin: 0;
+  padding-bottom: 0.5rem;
+  padding-top: 0.5rem;
+  text-align: center;
+`;
