+++ components/ImageBox/index.js
@@ -1,0 +1,3 @@
+import { ImageBox } from './styles';
+
+export default ImageBox;
