+++ components/MainContent/styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Container = styled('div')`
+  @media (max-width: 58.75rem) {
+    padding: 0 1.5rem;
+  }
+`;
+
+export const Main = styled('main')`
+  margin: auto;
+  max-width: 58.75rem;
+  padding: 3.875rem 0 6.25rem;
+  @media (max-width: 58.75rem) {
+    padding-top: 1.25rem;
+  }
+`;
